Row Component refactor
