+++ src/ui/BlurBox.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 
 import { gradients } from '../utils/theme'
 
-export const DimBlurBgBox = styled(Box)`
+export const BlurBox = styled(Box)`
   border-radius: 12px;
   border: 1px solid ${({ theme }) => theme.purple5};
